feat(web): add Kanban drag & drop with dnd-kit
- Enable issue drag & drop (intra-column reorder & cross-column move)
- Optimistic UI update; sync to server:
  - PATCH /issues/:id { columnId }
  - POST /columns/:id/reorder { issueIds }
- Add DndContext/Sensors and SortableContext per column
- Minor UI polish (cards, spacing)
