+++ apps/web/next.config.ts
@@ -1,11 +1,13 @@
 import type { NextConfig } from "next";
 
+const API_URL = process.env.API_URL ?? "http://127.0.0.1:3001";
+
 const nextConfig: NextConfig = {
     async rewrites() {
         return [
             {
                 source: "/api/:path*",
-                destination: "http://localhost:3001/:path*",
+                destination: `${API_URL}/:path*`,
             },
         ];
     },
